requirements.txt
flask
plotly
matplotlib
scipy
pandas
paho-mqtt
gunicorn
